Reorganize list components
