+++ crates/storybook2/src/story_selector.rs
@@ -21,6 +21,7 @@ pub enum ComponentStory {
     IconButton,
     Keybinding,
     Label,
+    List,
     ListItem,
     Scroll,
     Text,
@@ -40,6 +41,7 @@ impl ComponentStory {
             Self::IconButton => cx.build_view(|_| ui::IconButtonStory).into(),
             Self::Keybinding => cx.build_view(|_| ui::KeybindingStory).into(),
             Self::Label => cx.build_view(|_| ui::LabelStory).into(),
+            Self::List => cx.build_view(|_| ui::ListStory).into(),
             Self::ListItem => cx.build_view(|_| ui::ListItemStory).into(),
             Self::Scroll => ScrollStory::view(cx).into(),
             Self::Text => TextStory::view(cx).into(),
