+++ crates/ui2/src/components/stories.rs
@@ -6,7 +6,9 @@ mod icon;
 mod icon_button;
 mod keybinding;
 mod label;
+mod list;
 mod list_item;
+
 pub use avatar::*;
 pub use button::*;
 pub use checkbox::*;
@@ -15,4 +17,5 @@ pub use icon::*;
 pub use icon_button::*;
 pub use keybinding::*;
 pub use label::*;
+pub use list::*;
 pub use list_item::*;
